+++ edgy/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = "0.35.4"
+__version__ = "0.35.5"
 from typing import TYPE_CHECKING
 
 from ._monkay import Instance, create_monkay